Organise into Packages
Organise the code into packages dukes.engine, dukes.command, dukes.task and dukes.util
