+++ src/main/java/dukes/command/Command.java
@@ -1,4 +1,4 @@
-package dukes.engine;
+package dukes.command;
 
 import java.io.File;
 import java.io.FileNotFoundException;
