+++ src/main/java/dukes/command/DeleteCommand.java
@@ -1,6 +1,7 @@
-package dukes.engine;
+package dukes.command;
 
 import dukes.util.*;
+import dukes.task.*;
 
 import java.io.File;
 import java.io.FileNotFoundException;
