+++ src/main/java/dukes/util/UI.java
@@ -1,4 +1,4 @@
-package dukes.engine;
+package dukes.util;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -7,6 +7,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import dukes.util.*;
+import dukes.task.*;
 
 import java.time.format.DateTimeParseException;
 import java.util.*;
